[bugfix] Only index if there's stuff to index
A colleague found a fun edge case. He added three new empty files to a
project, and Expert wouldn't start. This is because of a divide by
zero error when there are no bytes to index, which causes progress
updates to fail (indexing completed successfully).

This fix checks for this condition and no-ops if there's nothing to index.